docs: :memo: README as qmd to include badges and citation (#1419)
# Description

Convert README to qmd so that we can insert the badges and the citation
information from the CITATION.cff file (plus a small edit to that file).

I also added a pre-render step to the quarto build to update the README
as well whenever the documentation is built.

Closes #1411

This PR needs a quick review.

## Checklist

- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>