fix: :bug: dedent `description` in `as_readme_text()` and `write_properties()` (#1476)
# Description

This PR adds an internal function `_to_dedented_text()` that's used in
`as_readme_text()` and `write_properties()` to dedent the
`PackageProperties.description` in case it's indented. This way, the
description text isn't indented in `datapackage.json` and `README.md`.

See issue below for example.

Closes #1475 

This PR needs an in-depth review.

## Checklist

- [X] Added or updated tests
- [ ] Ran `just run-all`

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>