docs: :memo: clarify intro and data constraints (#1585)
# Description

I think it's helpful to be more specific on what Sprout does already on
the intro page (without getting into too many details). I was inspired
by some of the language in
https://sprout.seedcase-project.org/docs/design/ and elaborated on it.

Let me know if I have done the commits correctly here; my understanding
is that since this is not a repo about only documentation, I should use
the `docs: :memo:` indicator for all docs-related commits.

Closes https://github.com/seedcase-project/seedcase-sprout/issues/1574
and starts to address
https://github.com/seedcase-project/seedcase-sprout/issues/1575

Needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [ ] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>