Remove the overview steps
I have a hard time following these myself
and its not clear to me that they are helpful
in understanding the rest of the material
(or exactly corresponding to it).
Some of them also rely on terms
that are only introduced later,
such as "batch".
It's possible that an updated version of these
can be added back later.